+++ packages/rollup-plugin-html/src/RollupPluginHTMLOptions.ts
@@ -41,6 +41,8 @@ export interface RollupPluginHTMLOptions {
   absolutePathPrefix?: string;
   /** When set to true, will insert meta tags for CSP and add script-src values for inline scripts by sha256-hashing the contents */
   strictCSPInlineScripts?: boolean;
+  /** Bundle assets reference from CSS via `url` */
+  bundleAssetsFromCss?: boolean;
 }
 
 export interface GeneratedBundle {